feat: AI审核功能完整实现 + 多项Bug修复
## AI审核 (Phase 1-2)
- 新建 DeepSeek API 审核服务
- 三层过滤架构: 规则引擎→AI语义→降级兜底
- SQL迁移: ai_audit_log/sensitive_words/ai_audit_config
- 管理界面: 统计/配置/敏感词/日志 4标签页

## Bug修复
- CSP: script-src-attr unsafe-inline
- 数据库编码 charset=utf8mb4
- 8个SQL文件INT→BIGINT UNSIGNED
- 限流/axios/ioredis/Pinia等多项修复
- 新建Login.vue登录页面

+++ backend/sql/29_debate_score.sql
@@ -1,7 +1,7 @@
 CREATE TABLE `debate_score` (
   `id` INT UNSIGNED NOT NULL AUTO_INCREMENT COMMENT '评分ID',
-  `topic_id` INT UNSIGNED NOT NULL COMMENT '关联辩论话题',
-  `voter_id` INT UNSIGNED NOT NULL COMMENT '评分者ID',
+  `topic_id` BIGINT UNSIGNED NOT NULL COMMENT '关联辩论话题',
+  `voter_id` BIGINT UNSIGNED NOT NULL COMMENT '评分者ID',
   `target_stance` TINYINT NOT NULL COMMENT '评分对象：1=正方, 0=反方',
   `criterion_key` VARCHAR(50) NOT NULL COMMENT '评分项标识（对应模板 scoring.criteria[].key）',
   `score` TINYINT UNSIGNED NOT NULL COMMENT '得分（1-10）',
